syncronize worker cleanup thread with its shutdown

+++ core/src/main/scala/org/apache/spark/deploy/worker/Worker.scala
@@ -72,6 +72,8 @@ private[deploy] class Worker(
   private val cleanupThreadExecutor = ExecutionContext.fromExecutorService(
     ThreadUtils.newDaemonSingleThreadExecutor("worker-cleanup-thread"))
 
+  private val cleanupLock = new Object
+
   // For worker and executor IDs
   private def createDateFormat = new SimpleDateFormat("yyyyMMddHHmmss", Locale.US)
   // Send a heartbeat every (heartbeat timeout) / 4 milliseconds
@@ -449,28 +451,32 @@ private[deploy] class Worker(
       // Spin up a separate thread (in a future) to do the dir cleanup; don't tie up worker
       // rpcEndpoint.
       // Copy ids so that it can be used in the cleanup thread.
-      val appIds = (executors.values.map(_.appId) ++ drivers.values.map(_.driverId)).toSet
-      val cleanupFuture = concurrent.Future {
-        val appDirs = workDir.listFiles()
-        if (appDirs == null) {
-          throw new IOException("ERROR: Failed to list files in " + appDirs)
-        }
-        appDirs.filter { dir =>
-          // the directory is used by an application - check that the application is not running
-          // when cleaning up
-          val appIdFromDir = dir.getName
-          val isAppStillRunning = appIds.contains(appIdFromDir)
-          dir.isDirectory && !isAppStillRunning &&
-          !Utils.doesDirectoryContainAnyNewFiles(dir, APP_DATA_RETENTION_SECONDS)
-        }.foreach { dir =>
-          logInfo(s"Removing directory: ${dir.getPath}")
-          Utils.deleteRecursively(dir)
-        }
-      }(cleanupThreadExecutor)
+      cleanupLock.synchronized {
+        if (!cleanupThreadExecutor.isShutdown) {
+          val appIds = (executors.values.map(_.appId) ++ drivers.values.map(_.driverId)).toSet
+          val cleanupFuture = concurrent.Future {
+            val appDirs = workDir.listFiles()
+            if (appDirs == null) {
+              throw new IOException("ERROR: Failed to list files in " + appDirs)
+            }
+            appDirs.filter { dir =>
+              // the directory is used by an application - check that the application is not running
+              // when cleaning up
+              val appIdFromDir = dir.getName
+              val isAppStillRunning = appIds.contains(appIdFromDir)
+              dir.isDirectory && !isAppStillRunning &&
+                !Utils.doesDirectoryContainAnyNewFiles(dir, APP_DATA_RETENTION_SECONDS)
+            }.foreach { dir =>
+              logInfo(s"Removing directory: ${dir.getPath}")
+              Utils.deleteRecursively(dir)
+            }
+          }(cleanupThreadExecutor)
 
-      cleanupFuture.failed.foreach(e =>
-        logError("App dir cleanup failed: " + e.getMessage, e)
-      )(cleanupThreadExecutor)
+          cleanupFuture.failed.foreach(e =>
+            logError("App dir cleanup failed: " + e.getMessage, e)
+          )(cleanupThreadExecutor)
+        }
+      }
 
     case MasterChanged(masterRef, masterWebUiUrl) =>
       logInfo("Master has changed, new master is at " + masterRef.address.toSparkURL)
@@ -666,7 +672,9 @@ private[deploy] class Worker(
   }
 
   override def onStop() {
-    cleanupThreadExecutor.shutdownNow()
+    cleanupLock.synchronized {
+      cleanupThreadExecutor.shutdownNow()
+    }
     metricsSystem.report()
     cancelLastRegistrationRetry()
     forwardMessageScheduler.shutdownNow()
